ingestion/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
boto3
requests
apache-airflow-providers-amazon
pytest
pytest-mock
moto[s3]
requests-mock
